requirements.txt
#must use djongo as 1.2.36
apipkg==1.5
asgiref==3.2.7
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==17.4.0
Automat==0.6.0
blinker==1.4
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
click==6.7
#cloud-init==19.3
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.3
cryptography==2.8
cssselect2==0.2.2
dataclasses==0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
Django==3.0.5
django-cors-headers==3.2.0
django-dynamic-model==0.1.0
django-extensions==2.2.8
django-filter==2.2.0
django-model-utils==4.0.0
django-rest-swagger==2.2.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djongo
EasyProcess==0.2.9
ec2-hibinit-agent==1.0.0
entrypoints==0.3
execnet==1.7.1
flake8==3.7.9
graphviz==0.13.2
gunicorn==20.0.4
hibagent==1.0.1
html5lib==1.0.1
httplib2==0.9.2
hyperlink==17.3.1
idna==2.8
importlib-metadata==1.6.0
incremental==16.10.1
itypes==1.1.0
Jinja2==2.10.3
joblib==0.14.0
jsonfield==2.0.2
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
netifaces==0.10.4
numpy==1.17.4
oauthlib==2.0.6
openapi-codec==1.3.2
packaging==20.3
PAM==0.4.2
pandas==0.25.3
pandas-compat==0.1.1
pbr==5.4.4
pdfkit==0.6.1
Pillow==7.0.0
pluggy==0.13.1
ply==3.11
psycopg2-binary==2.8.4
py==1.8.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.5.0
pycparser==2.19
pycrypto==2.6.1
pydot==1.4.1
pyflakes==2.1.1
Pygments==2.5.1
pygobject==3.26.1
PyHamcrest==2.0.2
PyJWT==1.5.3
pymongo==3.9.0
pyOpenSSL==17.5.0
pyparsing==2.4.6
PyPDF2==1.26.0
Pyphen==0.9.5
pyserial==3.4
pytest==5.3.5
pytest-cov==2.8.1
pytest-flake8==1.0.4
pytest-forked==1.1.3
pytest-pythonpath==0.7.3
pytest-sugar==0.9.2
pytest-xdist==1.31.0
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2019.3
PyVirtualDisplay==0.2.5
pywin32-ctypes==0.2.0
pyxdg==0.25
PyYAML==3.12
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.1.5
scikit-learn==0.22.2.post1
scipy==1.4.1
SecretStorage==2.3.1
service-identity==16.0.0
simplejson==3.17.0
six==1.13.0
sklearn==0.0
sqlparse
ssh-import-id==5.7
stevedore==1.31.0
systemd-python==234
termcolor==1.1.0
tinycss2==1.0.2
Twisted==17.9.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.0
urllib3==1.25.7
uWSGI==2.0.18
virtualenv==16.7.9
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wcwidth==0.1.8
WeasyPrint==51
webencodings==0.5.1
wkhtmltopdf==0.2
xhtml2pdf==0.2.4
xlrd==1.2.0
zipp==3.1.0
zope.interface==4.7.1
